Safari-609.1.11.1/LayoutTests/svg/animations/svgstring-animation-fallback-to-discrete.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
<script src="../dynamic-updates/resources/SVGTestCase.js"></script>
<script src="resources/SVGAnimationTestCase.js"></script>
</head>
<body onload="runSMILTest()">
<h1>SVG 1.1 dynamic animation tests</h1>
<p id="description"></p>
<div id="console"></div>
<script>
description("Tests fallback to calcMode='discrete' on animation of SVGString with 'values'.");
createSVGTestCase();

// Setup test document
var rect = createSVGElement("rect");
rect.setAttribute("id", "rect");
rect.setAttribute("width", "100");
rect.setAttribute("height", "100");
rect.setAttribute("fill", "green");
rect.setAttribute("onclick", "executeTest()");

var animate = createSVGElement("animate");
animate.setAttribute("id", "animation");
animate.setAttribute("attributeName", "visibility");
animate.setAttribute("begin", "click");
animate.setAttribute("dur", "6s");
animate.setAttribute("calcMode", "linear");
animate.setAttribute("values", "visible ; hidden ; visible");
rect.appendChild(animate);
rootSVGElement.appendChild(rect);

// Setup animation test
function sample1() {
    // Check initial/end conditions
    shouldBe("getComputedStyle(rect).visibility", "'visible'");
    shouldBeEqualToString("rect.style.visibility", "");
}

function sample2() {
    shouldBe("getComputedStyle(rect).visibility", "'hidden'");
    shouldBeEqualToString("rect.style.visibility", "");
}

function sample3() {
    shouldBe("getComputedStyle(rect).visibility", "'visible'");
    shouldBeEqualToString("rect.style.visibility", "");
}

function executeTest() {
    const expectedValues = [
        // [animationId, time, sampleCallback]
        ["animation", 1.999, sample1],
        ["animation", 2.001, sample2],
        ["animation", 5.999, sample3],
        ["animation", 6.001, sample1]
    ];

    runAnimationTest(expectedValues);
}

var successfullyParsed = true;
</script>
</body>
</html>
